completed part2.16-2.17
note: notification popups are buggy when user clicks on multiple actions. i believe this is because the notifications components keeps remounting before popup timer ends. because the exercise only asks to show various types of notification, i have decided that fixing this is error is out of the scope of this problem.